+++ crates/renzora_editor/src/plugins.rs
@@ -77,5 +77,4 @@ pub(crate) fn add_editor_plugins(app: &mut App) {
     app.add_plugins(renzora_update::UpdatePlugin::default());
     app.add_plugins(renzora_viewport::ViewportPlugin::default());
     app.add_plugins(renzora_volumetric_fog_editor::VolumetricFogEditorPlugin::default());
-    app.add_plugins(renzora_water_editor::WaterEditorPlugin::default());
 }